// src/tsco-cli/enums/class-member-type.ts
export enum ClassMemberType
{
    privateStaticReadOnlyProperties = "privateStaticReadOnlyProperties",
    privateReadOnlyProperties = "privateReadOnlyProperties",
    privateStaticProperties = "privateStaticProperties",
    privateProperties = "privateProperties",

    protectedStaticReadOnlyProperties = "protectedStaticReadOnlyProperties",
    protectedReadOnlyProperties = "protectedReadOnlyProperties",
    protectedStaticProperties = "protectedStaticProperties",
    protectedProperties = "protectedProperties",

    publicStaticReadOnlyProperties = "publicStaticReadOnlyProperties",
    publicReadOnlyProperties = "publicReadOnlyProperties",
    publicStaticProperties = "publicStaticProperties",
    publicProperties = "publicProperties",

    staticBlockDeclarations = "staticBlockDeclarations",
    constructors = "constructors",

    publicStaticAccessors = "publicStaticAccessors",
    publicAccessors = "publicAccessors",
    publicAbstractAccessors = "publicAbstractAccessors",

    privateStaticAccessors = "privateStaticAccessors",
    privateAccessors = "privateAccessors",

    protectedStaticAccessors = "protectedStaticAccessors",
    protectedAccessors = "protectedAccessors",
    protectedAbstractAccessors = "protectedAbstractAccessors",

    publicStaticGettersAndSetters = "publicStaticGettersAndSetters",
    publicGettersAndSetters = "publicGettersAndSetters",
    publicAbstractGettersAndSetters = "publicAbstractGettersAndSetters",

    privateStaticGettersAndSetters = "privateStaticGettersAndSetters",
    privateGettersAndSetters = "privateGettersAndSetters",

    protectedStaticGettersAndSetters = "protectedStaticGettersAndSetters",
    protectedGettersAndSetters = "protectedGettersAndSetters",
    protectedAbstractGettersAndSetters = "protectedAbstractGettersAndSetters",

    publicStaticMethods = "publicStaticMethods",
    publicMethods = "publicMethods",
    publicAbstractMethods = "publicAbstractMethods",

    protectedStaticMethods = "protectedStaticMethods",
    protectedMethods = "protectedMethods",
    protectedAbstractMethods = "protectedAbstractMethods",

    privateStaticMethods = "privateStaticMethods",
    privateMethods = "privateMethods"
}